Update FolderSelector in mobile app ItemNameField to support tree view (#1695)
